Move VersionedBlob to chia.util.streamable
