+++ chia/_tests/plotting/test_plot_manager.py
@@ -28,7 +28,7 @@ from chia.plotting.util import (
 from chia.simulator.block_tools import get_plot_dir
 from chia.util.config import create_default_chia_config, lock_and_load_config, save_config
 from chia.util.ints import uint16, uint32
-from chia.util.misc import VersionedBlob
+from chia.util.streamable import VersionedBlob
 
 log = logging.getLogger(__name__)
