+++ chia/_tests/wallet/test_wallet_coin_store.py
@@ -10,8 +10,8 @@ from chia._tests.util.db_connection import DBConnection
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.ints import uint8, uint16, uint32, uint64
-from chia.util.misc import UInt32Range, UInt64Range, VersionedBlob
-from chia.util.streamable import Streamable
+from chia.util.misc import UInt32Range, UInt64Range
+from chia.util.streamable import Streamable, VersionedBlob
 from chia.wallet.puzzles.clawback.metadata import ClawbackMetadata
 from chia.wallet.util.query_filter import AmountFilter, HashFilter
 from chia.wallet.util.wallet_types import CoinType, WalletType
